Visual Completo Painel Administrativo, menu+Usuario
Visual Completo Painel Administrativo, menu+Usuario

+++ EJB/SuperDia/ejbModule/br/com/sb/ClienteBean.java
@@ -2,9 +2,9 @@ package br.com.sb;
 
 import javax.ejb.Remote;
 import javax.ejb.Stateless;
-import javax.persistence.EntityExistsException;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.persistence.Query;
 
 import br.com.interfacebean.ICliente;
 import br.com.modelo.Usuario;
@@ -17,34 +17,42 @@ public class ClienteBean implements ICliente {
 	
 	@Override
 	public boolean autoAdiciona(Usuario usuario) {
-		try {
+		if(!entityExists(usuario)) {
 			usuario.setPerfil(Usuario.perfis.Cliente.name());
 			em.persist(usuario);
 			return true;
-		} catch (EntityExistsException e) {
-			return false;
 		}
+		return false;
 	}
 
 	@Override
 	public boolean autoAltera(Usuario usuario) {
-		try {
-			usuario.setPerfil(Usuario.perfis.Cliente.name());
+		if(entityExists(usuario)) {
 			em.merge(usuario);
 			return true;
-		} catch (IllegalArgumentException e) {
-			return false;
 		}
+		return false;
 	}
 
 	@Override
 	public boolean autoRemove(Usuario usuario) {
-		try {
+		if(entityExists(usuario)) {
 			em.remove(em.merge(usuario));
 			return true;
-		} catch (IllegalArgumentException e) {
+		}
+		return false;
+	}
+	
+	private boolean entityExists(Usuario usuario) {
+		try {
+			Query q = em.createQuery("SELECT u FROM Usuario u WHERE u.usuario = :usuario");
+			
+			q.setParameter("usuario", usuario.getUsuario());
+			
+			q.getSingleResult();
+			return true;
+		}catch (Exception e) {
 			return false;
 		}
 	}
-
 }
